Remove iocp receive pool and improve receive list
Removes the receive pool in favor of dynamically allocating everytime.
This is not particularly effecient and should be replaced in the future
but it is less convoluted than the previous linked list implementation.

Also improves the IO packet receive list by using `std::queue` instead
of the linked list method.

These improvements fix release builds crashing.