Replace login request packet with new flatbuffer
Replaces the client login request packet with the new packet class which
uses flatbuffers underneath. There is still some work to be done in
cleaning up the code on the server side to queue the packet for
processing after receiving it (sql thread). Also the old login
references can be cleaned up. Finally, it would appear the game server
is currently timing out or not connecting.