Add generate_launchers script
This script will generate convenience scripts to launch the servers and
client. It assumes some defaults which can be overridden. This replaces
the need to have hard-coded `.bat` files saved in the source tree and
generates the launchers in the parent dir in anticipation of separating
out the client files.