Fix inventory deletion bug
In the update character process, the order of insertions/updates was
such that existing data was updated and then we removed the "stale"
items. However, this would result valid items being deleted if they were
put into the same slot marked as stale. This update does the deletion
first then the insertion/updates.