Add support for server configurable base mspeed
This pull request improve the move speed data structures in the server
and lays out work for future values to be synced better. It also
refactors some of the move speed function commands for better clarity
and integrate the ability to configure the "base" move speed from the
server side toml config that gets applied to all players.

- Implement `CharacterStats` server side
- Add `/stats` gm command
- Replace server side invocations of `get_ori_mspeed` with
  `total_move_speed()`
- Integrate base move speed value from server toml config