Fix default items not using unique uuid
In the world server, when creating a new character, the default
items used for reference would use the same uuid every time. This would
result in a database conflict after the first character was ever
created and no new characters could be reated. This fix forces the
uuids to be reset for each new character.