Refactor move speed functions
This change removes most of the intermediate functions in the client
and server for move speed (e.g. GetOri_Move_Speed, Get_Move_Speed, etc.)
in favor of simpler access to this->stats.move_speed,
this->total_move_speed(), this->adjusted_move_speed (client), etc. It
also uses the CharacterStats structure in NPCs and MOBs.