Remove zone/npc/event var database code
This removes functionality related to saving zone economy data, npc data
and event data to/from the database. The reasoning is that this data was
essentially captured from the data files anyway, duplicating them into
the server was redundant. Furthermore, the actual implementation was
wonky and doesn't provide too much value. If there's a need for similar
functionality in the future (map/npc/event specific logic to be
controlled by the database) it should be built from scratch.