Fix bug where character with no union points could not be selected
If there was no associated row in the `union_points` table then the
existing query would fail. This changes the query so that default values
will be used in the case that there is no row (which will occur with
all new characters until they have been fully saved the first time).