Port messenger to postgres and fix existing bugs
Ported the messenger friends list to postgres with new tables. Replaced
the existing behavior of saving friends only on logout to updating the
database friends list as the add/delete is requested. An existing bug
regarding adding not working correctly was also fixed. The issue was in
the "select character" procedure whereby the character was added to the
global user list before the database id for that character was
populated. The adding of the character to the list triggered the
messenger login process too early resulting in it not having a valid id
to perform its login sequence.