Consolidate databases and migrations
Ultimately there is no need to split these databases because the chances
of running multiple game instances in multiple geolocations with one
central account database is low. The added overhead of this code is not
worth it and chances are a different system will be required if we ever
reach that stage (e.g. oauth2 auth with website vs. dedicated account
table)