Sync hp/mp recovery and increase rate
Removes the recovery rate calculations from the client and updates the
server to send hp/mp updates during the recovery making it
authoritative. Uses the existing stats packet to send this data.

Additionally this pull request introduces the beginning of a new
networking system that pulls the logic out of `classUSER`.