Fix character move mode sync and remove client calculations
Fixes a mismatch between server and client character move mode + move
speed by moving the calculations to the server. The move packet was
ported to flatbuffers and movespeed was added to the packet. This fixes
a bug where the mobs would walk at running speeds or run at walking
speeds.

Changes:
 - Client: Remove calculations for npc move speed based on move mode
 - Client: Add support for receiving new `CharacterMove` packet
 - Common: Add new `CharacterMove` flatbuffer packet
 - Common: Add new `Vec3` flatbuffer packet type
 - Common: Add new utils to help convert to/from packet types
 - GS: Fix bug in `ZoneSECTOR::GetSector` not getting the right sector
 - GS: Add new method to send flatbuffer packets to nearby members
 - GS: Add new base method `CObjAI::update_speed` with overrides for avatar
   and mob
 - GS: Update `Send_gsv_MOVE` to use new flatbuffer packets