Fix move and attack not syncing move mode/speed
Fixes a bug whereby characters would not have the correct move speed
after a move and attack command was initiated. This impact mobs who
would chase players using a running animation but sometimes with a
walking speed (if previously was walking). Now the move speed and move
mode is updated and sent with the move and attack packet.

- Client: Adding support to receive new `CharacterMoveAttack` packet
- GS: Added `CObjAI::can_attack` method to check if it's possible to
  attack the target
- GS: Updated `CObjChar::SetCMD_RUNnATTACK` to update movespeed/movemode
  and send new `CharacterMoveAttack` packet
- Common-lib: Add new `CharacterMoveAttack` packet and move
  `CharacterMoveMode` enum to its own FBS file