Update scripts after asset merge and disable CI
This updates the various build/dist/etc. scripts to work with the new
source tree that includes the assets. Main changes:

- bake.ps1 now bakes the assets to the `bin/assets` directory
- pack.ps1 default out directory changed to `dist/release/client`
- dist.ps1 copies all the files required for distribution to the `dist/`
  directory
- dist.ps1 now copies the individual migrations in addition to the
  squashed migrations
- squashed-migrations.ps1 was updated with the new single database
  format (previously had two databases, game and account)

CI was temporarily disabled since we don't have a build machine
available. It will be updated once a new machine is added and the
automated deploy functionality is developed.