Update README and add dev-setup script
The dev-setup script helps setup hardlinks and junctions or the client
and server data. This makes it easier to development and is better than
using the baked asset directory directly since it will prevent it from
being polluted by various files the server and client create at runtime.

The script creates folders in `dev/` that should be used as the working
directories for server and client during development.