Consolidate io_quest.h/cpp
This consolidates the io_quest.h and io_quest.cpp files from the client
and sho_gameserver projects. Where possible, common functionality was
replicated. Otherwise, the prepocessor was used heavily to merge the two
implementations. Additionally this is untested but fully compiles, the
runtime implications of these are yet to be determined.

WARNING: Likely buggy, needs testing.