Add discord sdk to client
Integrated the discord game sdk into the client. The client details are
configurable in discord_config.h with the default settings for ROSE
Next. Discord integraton can be disabled by unsetting the `DISCORD`
preprocessor variable.

The user's character name, job and level will be displayed when they are
in a map. If this changes it will only update when they change map. When
not in a map it shows no information.